sequence: Remove duplication

+++ egs/wsj/s5/steps/nnet3/train_discriminative.sh
@@ -51,7 +51,7 @@ num_threads=16  # this is the default but you may want to change it, e.g. to 1 i
                 # using GPUs.
 
 cleanup=true
-keep_model_iters=1
+keep_model_iters=100
 remove_egs=false
 src_model=  # will default to $degs_dir/final.mdl
 
@@ -194,7 +194,7 @@ if [ $stage -le -1 ]; then
 
   $cmd $dir/log/convert.log \
     nnet3-am-copy --learning-rate=$learning_rate "$src_model" $dir/0.mdl || exit 1;
-    
+
   ln -sf 0.mdl $dir/epoch0.mdl
 fi
 
@@ -310,47 +310,33 @@ while [ $x -lt $num_iters ]; do
       rm $dir/$[$x-1].mdl
     fi
 
-    if [ ! -z "${iter_to_epoch[$x]}" ]; then
-      e=${iter_to_epoch[$x]}
-      ln -sf $x.mdl $dir/epoch$e.mdl
-    fi
-      
-    if [ ! -z "${iter_to_epoch[$x]}" ]; then
-      (
-        e=${iter_to_epoch[$x]}
-        rm $dir/.error 2> /dev/null
-
-        steps/nnet3/adjust_priors.sh --egs-type degs \
-          --num-jobs-compute-prior $num_jobs_compute_prior \
-          --cmd "$cmd" --use-gpu false \
-          --use-raw-nnet false --iter epoch$e $dir $degs_dir \
-          || { touch $dir/.error; echo "Error in adjusting priors. See $dir/log/adjust_priors.epoch$e.log"; exit 1; }
-      ) &
-    fi
-
     [ -f $dir/.error ] && { echo "Found $dir/.error. Error on iteration $x"; exit 1; }
   fi
 
   rm $dir/cache.$x 2>/dev/null || true
   x=$[$x+1]
   num_archives_processed=$[num_archives_processed+num_jobs_nnet]
+
+  if [ $stage -le $x ] && [ ! -z "${iter_to_epoch[$x]}" ]; then
+    e=${iter_to_epoch[$x]}
+    ln -sf $x.mdl $dir/epoch$e.mdl
+
+    (
+      rm $dir/.error 2> /dev/null
+
+      steps/nnet3/adjust_priors.sh --egs-type degs \
+        --num-jobs-compute-prior $num_jobs_compute_prior \
+        --cmd "$cmd" --use-gpu false \
+        --minibatch-size $minibatch_size \
+        --use-raw-nnet false --iter epoch$e $dir $degs_dir \
+        || { touch $dir/.error; echo "Error in adjusting priors. See $dir/log/adjust_priors.epoch$e.log"; exit 1; }
+    ) &
+  fi
+
 done
 
 rm $dir/final.mdl 2>/dev/null
 cp $dir/$x.mdl $dir/final.mdl
-ln -sf final.mdl $dir/epoch$num_epochs_expanded.mdl
-
-if [ $stage -le $num_iters ]; then
-  steps/nnet3/adjust_priors.sh --egs-type degs \
-    --num-jobs-compute-prior $num_jobs_compute_prior \
-    --cmd "$cmd $prior_queue_opt" --use-gpu $use_gpu \
-    --minibatch-size $minibatch_size \
-    --use-raw-nnet false --iter epoch$num_epochs_expanded \
-    $dir $degs_dir || exit 1
-fi
-
-echo Done
-
 
 # function to remove egs that might be soft links.
 remove () { for x in $*; do [ -L $x ] && rm $(readlink -f $x); rm $x; done }
@@ -371,3 +357,8 @@ if $cleanup; then
     fi
   done
 fi
+
+wait
+[ -f $dir/.error ] && { echo "Found $dir/.error."; exit 1; }
+
+echo Done && exit 0
